techpack: dsp: Force AVCS CVP version 1 on msm8976/56
Force the AVCS version to CVP_VERSION_1 for all of the AVCS
services on msm8976/56: it cannot execute the Q6 Service Version
command on its Q6 DSP because of older firmware not supporting it.

We know that for this SoC the version is always 1, hence
forcing it is totally safe.

Change-Id: Ib4841bd2eefd31bb320403e77d26f66136927c5f